Change type GC_collect_a_little_inner argument from int to size_t (refactoring)
* alloc.c (GC_deficit): Change type from int to size_t.
* alloc.c (GC_rate): Change type from int to unsigned; reformat
comment.
* alloc.c (GC_rate): Rename n to n_blocks in comment.
* include/private/gc_priv.h (GC_collect_a_little_inner): Likewise.
* alloc.c (GC_set_rate): Cast value to unsigned.
* alloc.c (GC_get_rate): Cast GC_rate to int.
* alloc.c (GC_collect_a_little_inner): Rename n to n_blocks argument
and change its type from int to size_t.
* include/private/gc_priv.h (GC_collect_a_little_inner): Likewise.
* alloc.c (GC_collect_a_little_inner): Change type of i, max_deficit
local variables from int to size_t; ensure an underflow cannot happen
in GC_deficit-max_deficit.
* alloc.c (GC_stopped_mark): Change type of abandoned_at local
variable from int to unsigned; change print format specifier for
abandoned_at from %d to %u; count loop iterations in abandoned_at
starting from 1 instead of 0 (and use 0 as a marker that the collection
is abandoned); adjust code indentation.
* alloc.c (GC_expand_hp): Change type of n_blocks local variable from
word to size_t.
* malloc.c (GC_alloc_large): Do not cast n_blocks to int.

+++ malloc.c
@@ -60,7 +60,7 @@ STATIC ptr_t GC_alloc_large(size_t lb_adjusted, int k, unsigned flags,
     /* Do our share of marking work.    */
     if (GC_incremental && !GC_dont_gc) {
             ENTER_GC();
-            GC_collect_a_little_inner((int)n_blocks);
+            GC_collect_a_little_inner(n_blocks);
             EXIT_GC();
     }
